requirements.txt
blinker==1.4
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Dance==1.3.0
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
Flask-Migrate
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
lazy==1.4
MarkupSafe==1.1.0
oauthlib==3.0.1
requests==2.21.0
requests-oauthlib==1.2.0
signals==0.0.2
six==1.12.0
SQLAlchemy==1.2.18
SQLAlchemy-Utils==0.33.11
urllib3==1.24.1
URLObject==2.4.3
Werkzeug==0.14.1
python-dotenv
pyopenssl
